80 character limit
This makes it easier to write and read on both narrow and wide screens,
and makes it much much easier to do diffs between lines.

e.g., if I change one word in a very long line, it's not very easy to
see what changed.  But if it's changed in a shorter line, it's easier to
see.

See also:

https://www.smashingmagazine.com/2014/09/balancing-line-length-font-size-responsive-web-design/

And probably some more resources.